fix: clarify behavior of update() method on Endpoint
The update() method on the endpoint incorrectly documented what the
return values were. The code examples didn't expose this because in the
incorrect case the return was not used. Updated the documentation to
reflect the behavior but also rewrote the code to narrow down the
possibilities of what can be supplied so that there are narrower types
allowing more type checking in the future to be added. Maintained the
same behavior as the existing code with the exception of the tests which
now do not pass pylint checks and another where the error gets caught
for validity sooner than the code used to catch it.
Fixes #307